Handles use configure()
