+++ src/main.rs
@@ -10,19 +10,17 @@ mod template;
 mod common;
 
 use actix_web::middleware::Logger;
-use actix_web::{get,post,patch,delete,web,App, HttpResponse,HttpServer, Responder};
+use actix_web::{get,post,web,App, HttpResponse,HttpServer, Responder};
 
-use log::error;
-use tmflib::tmf620::product_offering::ProductOffering;
-use tmflib::tmf620::product_offering_price::ProductOfferingPrice;
+use model::tmf::tmf620::config_tmf620;
+use model::tmf::tmf622::config_tmf622;
+use model::tmf::tmf629::config_tmf629;
+use model::tmf::tmf632::config_tmf632;
+use model::tmf::tmf648::config_tmf648;
+use model::tmf::tmf674::config_tmf674;
+
 #[cfg(feature = "tmf648_v4")]
 use tmflib::tmf648::quote::Quote;
-#[cfg(feature = "tmf674_v4")]
-use tmflib::tmf674::geographic_site_v4::GeographicSite;
-#[cfg(feature = "tmf674_v5")]
-use tmflib::tmf674::geographic_site_v5::GeographicSite;
-
-use crate::model::tmf::tmf620::config_tmf620;
 
 use std::sync::Mutex;
 
@@ -34,15 +32,11 @@ use common::persist::Persistence;
 
 // TMFLIB
 use common::config::Config;
-use common::error::PlatypusError;
-use tmflib::tmf620::catalog::Catalog;
-use tmflib::tmf620::category::Category;
-use tmflib::tmf620::product_specification::ProductSpecification;
-use tmflib::tmf632::individual::Individual;
-use tmflib::tmf632::organization::Organization;
+
+
 use tmflib::tmf629::customer::Customer;
 use tmflib::tmf629::customer::CUST_STATUS;
-use tmflib::{HasId, HasLastUpdate};
+use tmflib::HasId;
 
 #[cfg(feature = "composable")]
 use crate::model::component::*;
@@ -51,9 +45,6 @@ use crate::template::*;
 
 //use crate::template::product::ProductTemplate;
 //use crate::model::component::product::ProductComponent;
-use crate::model::tmf::tmf632_party_management::TMF632PartyManagement;
-use crate::model::tmf::tmf620::tmf620_catalog_management::TMF620CatalogManagement;
-use crate::model::tmf::tmf674_geographic_site::TMF674GeographicSiteManagement;
 
 /// Fields for filtering output
 #[derive(Clone, Debug, Deserialize)]
@@ -64,226 +55,6 @@ pub struct QueryOptions {
     offset : Option<u16>,
     /// Filter on name
     name : Option<String>,
-}
-
-/// Get a list
-#[get("/tmf-api/productCatalogManagement/v4/{object}")]
-pub async fn tmf620_list_handler(
-    path : web::Path<String>,
-    tmf620: web::Data<Mutex<TMF620CatalogManagement>>,
-    persist: web::Data<Mutex<Persistence>>,
-    query : web::Query<QueryOptions>,
-) -> impl Responder {
-    let object = path.into_inner();
-    let query_opts = query.into_inner();
-    let persist = persist.lock().unwrap();
-    // Now have to pass persistence into tmf module here
-    tmf620.lock().unwrap().persist(persist.clone());
-
-    match object.as_str() {
-        "catalog" => {
-            let output = tmf620.lock().unwrap().get_catalogs(query_opts).await;
-            match output {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),
-            }
-        },
-        "category" => {
-            let output = tmf620.lock().unwrap().get_categories(query_opts).await;
-            match output {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),
-            }
-        },
-        "productSpecification" => {
-            let output = tmf620.lock().unwrap().get_specifications(query_opts).await;
-            match output {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),
-            }
-        },
-        "productOffering" => {
-            let output = tmf620.lock().unwrap().get_offers(query_opts).await;
-            match output {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),
-            }
-        }
-        "productOfferingPrice" => {
-            let output = tmf620.lock().unwrap().get_prices(query_opts).await;
-            match output {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),
-            }    
-        },
-        "importJob" => {
-            HttpResponse::BadRequest().json(PlatypusError::from("importJob: Not implemented"))
-        },
-        "exportJob" => {
-            HttpResponse::BadRequest().json(PlatypusError::from("exportJob: Not implemented"))
-        },
-        "hub" => {
-            HttpResponse::BadRequest().json(PlatypusError::from("Hub: Not implemented"))
-        },
-        "listener" => {
-            HttpResponse::BadRequest().json(PlatypusError::from("listener: Not implemented"))
-        },
-        _ => HttpResponse::BadRequest().json(PlatypusError::from("Bad Object: {object}")),
-    }
-}
-
-/// Get a specific object
-#[get("/tmf-api/productCatalogManagement/v4/{object}/{id}")]
-pub async fn tmf620_get_handler(
-    path : web::Path<(String,String)>,
-    tmf620: web::Data<Mutex<TMF620CatalogManagement>>,
-    query : web::Query<QueryOptions>,
-) -> impl Responder {
-    let (object,id) = path.into_inner();
-    let query_opts = query.into_inner();
-    
-    match object.as_str() {
-        "catalog" => {
-            let output = tmf620.lock().unwrap().get_catalog(id,query_opts).await;
-            match output {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),
-            }
-        },
-        "category" => {
-            let output = tmf620.lock().unwrap().get_category(id,query_opts).await;
-            match output {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),
-            }
-        },
-        "productSpecification" => {
-            let data = tmf620.lock().unwrap().get_specification(id,query_opts).await;
-            match data {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),    
-            }
-        },
-        "productOffering" => {
-            let data = tmf620.lock().unwrap().get_offer(id,query_opts).await;
-            match data {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),    
-            }
-        },
-        "productOfferingPrice" => {
-            let data = tmf620.lock().unwrap().get_price(id,query_opts).await;
-            match data {
-                Ok(o) => HttpResponse::Ok().json(o),
-                Err(e) => HttpResponse::InternalServerError().json(e),    
-            }
-        },
-        "importJob" => {
-            HttpResponse::BadRequest().json(PlatypusError::from("importJob: Not implemented"))
-        },
-        "exportJob" => {
-            HttpResponse::BadRequest().json(PlatypusError::from("exportJob: Not implemented"))
-        },
-        _ => HttpResponse::BadRequest().json(PlatypusError::from("Invalid Object: {object}"))
-    }
-}
-
-/// Update an object
-#[patch("/tmf-api/productCatalogManagement/v4/{object}/{id}")]
-pub async fn tmf620_patch_handler(
-    path : web::Path<(String,String)>,
-    raw: web::Bytes,
-    tmf620: web::Data<Mutex<TMF620CatalogManagement>>
-) -> impl Responder {
-    let (object,id) = path.into_inner();
-    let json = String::from_utf8(raw.to_vec()).unwrap();
-    match object.as_str() {
-        "productSpecification" => {
-            match tmf620.lock().unwrap().patch_specification(id,json).await {
-                Ok(r) => HttpResponse::Ok().json(r),
-                Err(e) => {
-                    error!("Could not delete: {e}");
-                    HttpResponse::BadRequest().json(e)
-                },
-            }
-        },
-        "productOffering" => {
-            match tmf620.lock().unwrap().patch_offering(id,json).await {
-                Ok(r) => HttpResponse::Ok().json(r),
-                Err(e) => {
-                    error!("Could not delete: {e}");
-                    HttpResponse::BadRequest().json(PlatypusError::from("PATCH: Bad object"))
-                },
-            }
-        },
-        "productOfferingPrice"  => {
-            match tmf620.lock().unwrap().patch_price(id,json).await {
-                Ok(r) => HttpResponse::Ok().json(r),
-                Err(e) => {
-                    error!("Could not delete: {e}");
-                    HttpResponse::BadRequest().json(PlatypusError::from("PATCH: Bad object"))
-                },
-            }
-        },
-        _ => HttpResponse::BadRequest().json(PlatypusError::from("PATCH: Bad object: {object}"))
-    } 
-}
-
-/// Detele an object
-#[delete("/tmf-api/productCatalogManagement/v4/{object}/{id}")]
-pub async fn tmf620_delete_handler(
-    path : web::Path<(String,String)>,
-    tmf620: web::Data<Mutex<TMF620CatalogManagement>>
-) -> impl Responder {
-    let (object,id) = path.into_inner();
-    match object.as_str() {
-        "catalog" => {
-            match tmf620.lock().unwrap().delete_catalog(id).await {
-                Ok(_b) => HttpResponse::NoContent(),
-                Err(e) => {
-                    error!("Could not delete: {e}");
-                    HttpResponse::BadRequest()
-                },     
-            }    
-        },
-        "category" => {
-            match tmf620.lock().unwrap().delete_category(id).await {
-                Ok(_b) => HttpResponse::NoContent(),
-                Err(e) => {
-                    error!("Could not delete: {e}");
-                    HttpResponse::BadRequest()
-                },     
-            }    
-        },
-        "productSpecification" => {
-            match tmf620.lock().unwrap().delete_specification(id).await {
-                Ok(_b) => HttpResponse::NoContent(),
-                Err(e) => {
-                    error!("Could not delete: {e}");
-                    HttpResponse::BadRequest()
-                },
-            }
-        },
-        "productOffering" => {
-            match tmf620.lock().unwrap().delete_offering(id).await {
-                Ok(_b) => HttpResponse::NoContent(),
-                Err(e) => {
-                    error!("Could not delete: {e}");
-                    HttpResponse::BadRequest()
-                },
-            }
-        },
-        "productOfferingPrice"  => {
-            match tmf620.lock().unwrap().delete_price(id).await {
-                Ok(_b) => HttpResponse::NoContent(),
-                Err(e) => {
-                    error!("Could not delete: {e}");
-                    HttpResponse::BadRequest()
-                },
-            }
-        },
-        _ => HttpResponse::BadRequest(),
-    }  
 }
 
 #[post("/tmflib/tmf629/customer")]
@@ -306,87 +77,6 @@ pub async fn tmf629_get_handler(
     HttpResponse::Ok()
 }
 
-#[get("/tmflib/tmf632/{object}")]
-pub async fn tmf632_list_handler(
-    path : web::Path<String>,
-    tmf632: web::Data<Mutex<TMF632PartyManagement>>,
-    query : web::Query<QueryOptions>,
-) -> impl Responder {
-    let query_opts = query.into_inner();
-    match path.as_str() {
-        "individual" => {
-            let result = tmf632.lock().unwrap().get_individuals(query_opts).await;
-            match result {
-                Ok(v) => HttpResponse::Ok().json(v),
-                Err(e) => HttpResponse::BadRequest().json(e),
-            }   
-        },
-        "organization" => todo!(),
-        _ => HttpResponse::BadRequest().json(PlatypusError::from("TMF632: Invalid Object"))
-    }  
-}
-
-/// Get a specific object
-#[get("/tmf-api/partyManagement/v4/{object}/{id}")]
-pub async fn tmf632_get_handler(
-    path : web::Path<(String,String)>,
-    tmf632: web::Data<Mutex<TMF632PartyManagement>>,
-    query : web::Query<QueryOptions>,
-) -> impl Responder {
-    let (object,id) = path.into_inner();
-    let query_opts = query.into_inner();
-    match object.as_str() {
-        "individual" => {
-            let result = tmf632.lock().unwrap().get_individual(id,query_opts).await;
-            match result {
-                Ok(v) => HttpResponse::Ok().json(v),
-                Err(e) => HttpResponse::BadRequest().json(e),
-            }       
-        },
-        _ => HttpResponse::BadRequest().json(PlatypusError::from("TMF632: Invalid Object"))    
-    }
-}
-
-#[post("/tmflib/tmf632/{object}")]
-pub async fn tmf632_post_handler(
-    path : web::Path<String>,
-    raw: web::Bytes,
-    tmf632: web::Data<Mutex<TMF632PartyManagement>>,
-) -> impl Responder {
-    let object = path.into_inner();
-    let json = String::from_utf8(raw.to_vec()).unwrap();
-    match object.as_str() {
-        "individual" => {
-            // Create individual object
-            let mut individual : Individual = serde_json::from_str(json.as_str()).unwrap();
-            individual.generate_id();
-            let records = tmf632.lock().unwrap().add_individual(individual.clone()).await;
-            match records {
-                Ok(r) => HttpResponse::Ok().json(r),
-                Err(e) => HttpResponse::BadRequest().json(e),
-            } 
-        },
-        "organization" => {
-            let mut organization : Organization = serde_json::from_str(json.as_str()).unwrap();
-            organization.generate_id();
-            HttpResponse::Ok().json(organization)
-        }
-        _ => {
-            HttpResponse::BadRequest().json(PlatypusError::from("TMF632: Invalid Object"))
-        }
-    } 
-}
-
-#[post("/tmflib/tmf648/quote")]
-pub async fn tmf648_create_handler(
-    body : web::Json<Quote>
-) -> impl Responder {
-    let data = body.into_inner();
-    HttpResponse::Ok().json(data)
-}
-
-
-
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
     let pkg = env!("CARGO_PKG_NAME");
@@ -396,11 +86,8 @@ async fn main() -> std::io::Result<()> {
 
     info!("Starting {pkg} v{ver}");
 
+    // Data objects to be pass in
     let persist = Persistence::new().await;
-    let tmf620 = TMF620CatalogManagement::new(persist.clone());
-    let tmf632 = TMF632PartyManagement::new(persist.clone());
-    let tmf674 = TMF674GeographicSiteManagement::new(persist.clone());
-
     let config = Config::new();
 
     // Extract port crom config, default if not found
@@ -409,21 +96,17 @@ async fn main() -> std::io::Result<()> {
    
     HttpServer::new(move || {
         App::new()
+            // Using the new configure() approach, we cannot pass persis in as
+            // configure() does not take additional arguments
             .app_data(web::Data::new(Mutex::new(persist.clone())))
-            .app_data(web::Data::new(Mutex::new(tmf632.clone())))
-            .app_data(web::Data::new(Mutex::new(tmf674.clone())))
             .app_data(web::Data::new(Mutex::new(config.clone())))
+            // New simple config functions.
             .configure(config_tmf620)
-            .service(tmf620_list_handler)
-            .service(tmf620_get_handler)
-            .service(tmf620_patch_handler)
-            .service(tmf620_delete_handler)
-            .service(tmf674_post_handler)
-            .service(tmf674_list_handler)
-            .service(tmf674_get_handler)
-            .service(tmf632_post_handler)
-            .service(tmf632_list_handler)
-            .service(tmf632_get_handler)
+            .configure(config_tmf622)
+            .configure(config_tmf629)
+            .configure(config_tmf648)
+            .configure(config_tmf632)
+            .configure(config_tmf674)
             .wrap(Logger::default())
     })
         .bind(("0.0.0.0",port))?
