+++ src/model/tmf/tmf629/tmf629_customer_management.rs
@@ -1,0 +1,10 @@
+//! Customer Management Module
+
+
+pub struct TMF629CustomerManagement {}
+
+impl TMF629CustomerManagement {
+    pub fn new() -> TMF629CustomerManagement {
+        TMF629CustomerManagement {}
+    }
+}
