+++ src/model/tmf/tmf648/tmf648_quote_management.rs
@@ -1,0 +1,10 @@
+//! Quote Management Module
+
+
+pub struct TMF648QuoteManagement {}
+
+impl TMF648QuoteManagement {
+    pub fn new() -> TMF648QuoteManagement {
+        TMF648QuoteManagement {}
+    }
+}
